documentbasedchatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain
langchain-community
langchain-postgres
langchain-huggingface
langchain-google-genai
google-generativeai
sentence-transformers
psycopg[binary]
pgvector
sqlalchemy
python-dotenv
elevenlabs
requests
num2words
supabase
python-docx
pypdf
beautifulsoup4
